Remove `code` and `data` directories under `Dev`
I was using this structure in order to keep data backups outside the
project tree and thus unknown to Git. It was a bit of a pain to manage,
as it's much more convenient to have all files pertinent to the project
within the project directory.

In the future I'll prefer creation of a `.data` directory underneath the
project root, and addition of that directory to the `.git/info/exclude`
file of the project.